refactor: extract CreatePolicies
